requirements.txt
playwright==1.38.0
pytest==7.4.0
pytest-playwright==0.3.0
allure-pytest==2.13.1
